+++ src/heuristica/testComplejidad.cpp
@@ -24,7 +24,7 @@ int main(int argc, const char* argv[]) {
         std::vector<int> sumaTiemposPorK(MAX_K / K_STEP);
         for (int instancia = 1; instancia <= CANT_INSTANCIAS; instancia++) {
             int n, m, u, v, original_k;
-            float w;
+            double w;
             std::cin >> n >> m >> original_k;
             Grafo g(n);
             for (int i = 0; i < m; i++) {
